renaming files
Files that export a class are capitalized to match their respective
class name. util.js is renamed to utils.js to match its folder name.
