+++ lib/Config.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const toml = require('toml');
-const util = require('./utils/util');
+const util = require('./utils/utils');
 
 
 class Config {
