+++ lib/Logger.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const winston = require('winston');
-const util = require('./utils/util');
+const util = require('./utils/utils');
 
 class Logger {
   constructor({ level, logDir, context }) {
