+++ lib/XUnion.js
@@ -1,11 +1,11 @@
 const Logger = require('./Logger');
-const Config = require('./config');
-const DB = require('./db/db');
-const OrderBook = require('./orderbook/orderbook');
-const LndClient = require('./lndclient/lndclient');
-const RpcServer = require('./rpcserver/rpcserver');
-const P2P = require('./p2p/p2p');
-const P2PServer = require('./p2p/p2pserver');
+const Config = require('./Config');
+const DB = require('./db/DB');
+const OrderBook = require('./orderbook/OrderBook');
+const LndClient = require('./lndclient/LndClient');
+const RpcServer = require('./rpcserver/RpcServer');
+const P2P = require('./p2p/P2P');
+const P2PServer = require('./p2p/P2PServer');
 
 class XUnion {
   constructor() {
