+++ lib/lndclient/LndClient.js
