+++ lib/p2p/P2P.js
@@ -1,7 +1,7 @@
 const net = require('net');
 const Logger = require('../Logger');
 const assert = require('assert');
-const OrderBook = require('../orderbook/orderbook');
+const OrderBook = require('../orderbook/OrderBook');
 
 class P2P {
   constructor(orderBook) {
