+++ lib/utils/utils.js
@@ -1,18 +1,18 @@
-const util = exports;
+const utils = exports;
 
-util.copyObject = (source, dest) => {
+utils.copyObject = (source, dest) => {
   const sourceKeys = Object.keys(source);
   for (let n = 0; n < sourceKeys.length; n += 1) {
     const key = sourceKeys[n];
     if (typeof dest[key] === 'object' && !Array.isArray(dest[key])) {
-      util.copyObject(source[key], dest[key]);
+      utils.copyObject(source[key], dest[key]);
     } else {
       dest[key] = source[key]; // eslint-disable-line no-param-reassign
     }
   }
 };
 
-util.nestObject = (obj) => {
+utils.nestObject = (obj) => {
   const nestedObject = {};
   const keys = Object.keys(obj);
   for (let n = 0; n < keys.length; n += 1) {
@@ -32,4 +32,4 @@ util.nestObject = (obj) => {
   return nestedObject;
 };
 
-util.getTsString = () => (new Date()).toLocaleString();
+utils.getTsString = () => (new Date()).toLocaleString();
