+++ lib/xuclient/XUClient.js
